move some func to utils.py
