+++ pyzoo/zoo/chronos/data/utils/utils.py
@@ -1,0 +1,40 @@
+#
+# Copyright 2018 Analytics Zoo Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+
+def _to_list(item, name, expect_type=str):
+    if isinstance(item, list):
+        return item
+    if item is None:
+        return []
+    _check_type(item, name, expect_type)
+    return [item]
+
+
+def _check_type(item, name, expect_type):
+    assert isinstance(item, expect_type), \
+        f"a {str(expect_type)} is expected for {name} but found {type(item)}"
+
+
+def _check_col_within(df, col_name):
+    assert col_name in df.columns, \
+        f"{col_name} is expected in dataframe while not found"
+
+
+def _check_col_no_na(df, col_name):
+    _check_col_within(df, col_name)
+    assert df[col_name].isna().sum() == 0, \
+        f"{col_name} column should not have N/A."
